requirements.txt
python-telegram-bot==21.5
python-dotenv==1.2.1
scikit-learn==1.7.2
numpy==2.3.4